+++ test/distributed/rpc/tensorpipe/test_ddp_under_dist_autograd.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 import unittest
 
+from torch.testing._internal.common_distributed import MultiProcessTestCase
 from torch.testing._internal.common_utils import TEST_WITH_ASAN, run_tests
 from torch.testing._internal.distributed.rpc.tensorpipe_rpc_agent_test_fixture import (
     TensorPipeRpcAgentTestFixture,
@@ -8,13 +9,23 @@ from torch.testing._internal.distributed.rpc.tensorpipe_rpc_agent_test_fixture i
 from torch.testing._internal.dist_utils import (
     dist_init,
 )
-from torch.testing._internal.distributed import ddp_under_dist_autograd_test
+from torch.testing._internal.distributed.ddp_under_dist_autograd_test import (
+    TestDdpComparison,
+    TestDdpUnderDistAutograd,
+)
 import torch.distributed.rpc as rpc
 
 @unittest.skipIf(
     TEST_WITH_ASAN, "Skip ASAN as torch + multiprocessing spawn have known issues"
 )
-class TestDdpUnderDistAutogradTensorPipe(TensorPipeRpcAgentTestFixture, ddp_under_dist_autograd_test.TestDdpUnderDistAutograd):
+class TestDdpUnderDistAutogradTensorPipe(
+    TestDdpUnderDistAutograd,
+    TensorPipeRpcAgentTestFixture,
+    MultiProcessTestCase,
+):
+    def setUp(self):
+        super().setUp()
+        self._spawn_processes()
 
     @dist_init
     def test_verify_backend_options(self):
@@ -23,7 +34,14 @@ class TestDdpUnderDistAutogradTensorPipe(TensorPipeRpcAgentTestFixture, ddp_unde
 @unittest.skipIf(
     TEST_WITH_ASAN, "Skip ASAN as torch + multiprocessing spawn have known issues"
 )
-class TestDdpComparisonTensorPipe(TensorPipeRpcAgentTestFixture, ddp_under_dist_autograd_test.TestDdpComparison):
+class TestDdpComparisonTensorPipe(
+    TestDdpComparison,
+    TensorPipeRpcAgentTestFixture,
+    MultiProcessTestCase,
+):
+    def setUp(self):
+        super().setUp()
+        self._spawn_processes()
 
     @dist_init
     def test_verify_backend_options(self):
