+++ test/distributed/test_ddp_under_dist_autograd.py
@@ -1,15 +1,28 @@
 #!/usr/bin/env python3
+import unittest
 
-from torch.testing._internal.distributed import ddp_under_dist_autograd_test
-from torch.testing._internal.common_utils import (
-    run_tests,
+from torch.testing._internal.common_distributed import MultiProcessTestCase
+from torch.testing._internal.distributed.ddp_under_dist_autograd_test import (
+    TestDdpComparison,
+    TestDdpUnderDistAutograd,
 )
+from torch.testing._internal.common_utils import TEST_WITH_ASAN, run_tests
 
-class TestDdpUnderDistAutogradWrapper(ddp_under_dist_autograd_test.TestDdpUnderDistAutograd):
-    pass
+@unittest.skipIf(
+    TEST_WITH_ASAN, "Skip ASAN as torch + multiprocessing spawn have known issues"
+)
+class TestDdpUnderDistAutogradWrapper(TestDdpUnderDistAutograd, MultiProcessTestCase):
+    def setUp(self):
+        super().setUp()
+        self._spawn_processes()
 
-class TestDdpComparison(ddp_under_dist_autograd_test.TestDdpComparison):
-    pass
+@unittest.skipIf(
+    TEST_WITH_ASAN, "Skip ASAN as torch + multiprocessing spawn have known issues"
+)
+class TestDdpComparison(TestDdpComparison, MultiProcessTestCase):
+    def setUp(self):
+        super().setUp()
+        self._spawn_processes()
 
 if __name__ == "__main__":
     run_tests()
